requirements.txt
discord.py>=2.3.2
aiosqlite>=0.19.0
python-dotenv>=1.0.1
pytest>=7.4.4
pytest-asyncio>=0.23.2
